feat(mtom): cleaning up PR
Refs: #1667
