+++ packages/core/src/external/carequality/ihe-gateway-v2/outbound/__tests__/process-dr.test.ts
@@ -1,13 +1,123 @@
 import fs from "fs";
 import path from "path";
 import { v4 as uuidv4 } from "uuid";
-import { OutboundDocumentRetrievalResp } from "@metriport/ihe-gateway-sdk";
 import { processDrResponse } from "../xca/process/dr-response";
-import { outboundDrRequest, testFiles, testFilesForUploadVerification } from "./constants";
+import { outboundDrRequest, testFiles, outboundDrRequestMtom } from "./constants";
 import { S3Utils } from "../../../../aws/s3";
-import { Config } from "../../../../../util/config";
-import { prepateMtomAttachments } from "./mtom";
+import { createMtomMessageWithAttachments, createMtomMessageWithoutAttachments } from "./mtom";
 
+describe("processDRResponse for MTOM with/without attachments and for different file types", () => {
+  beforeEach(() => {
+    jest.spyOn(S3Utils.prototype, "uploadFile").mockImplementation(() => {
+      return Promise.resolve({
+        Location: "http://example.com/mockurl",
+        ETag: '"mockedetag"',
+        Bucket: "mockedbucket",
+        Key: "mockedkey",
+      });
+    });
+
+    jest.spyOn(S3Utils.prototype, "getFileInfoFromS3").mockImplementation(() =>
+      Promise.resolve({
+        exists: false,
+      })
+    );
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  testFiles.forEach(({ name, mimeType, fileExtension }) => {
+    it(`[mtom without attachments]: should process the ${fileExtension} DR response correctly`, async () => {
+      const filePath = path.join(__dirname, "./files/", name);
+      const fileBuffer = fs.readFileSync(filePath);
+
+      const attachmentsData = [{ payload: fileBuffer, mimeType: mimeType }];
+      const mtomAttachments = await createMtomMessageWithAttachments(attachmentsData);
+      const response = await processDrResponse({
+        drResponse: {
+          mtomResponse: mtomAttachments,
+          gateway: outboundDrRequestMtom.gateway,
+          outboundRequest: outboundDrRequestMtom,
+        },
+      });
+
+      expect(response).toBeDefined();
+      expect(response.documentReference).toHaveLength(1);
+      expect(response.documentReference?.[0]?.contentType).toBe(mimeType);
+    });
+  });
+
+  testFiles.forEach(({ name, mimeType, fileExtension }) => {
+    it(`[mtom with attachments]: should process the ${fileExtension} DR response correctly`, async () => {
+      const xmlTemplatePath = path.join(__dirname, "./xmls/dr-no-mime-type.xml");
+      const xmlTemplate = fs.readFileSync(xmlTemplatePath, "utf8");
+
+      const fileContent = fs.readFileSync(path.join(__dirname, `./files/${name}`));
+      const fileContentB64 = fileContent.toString("base64");
+      const modifiedXml = xmlTemplate.replace(
+        "<Document></Document>",
+        `<Document>${fileContentB64}</Document>`
+      );
+
+      const mtomResponse = await createMtomMessageWithoutAttachments(modifiedXml);
+      const response = await processDrResponse({
+        drResponse: {
+          mtomResponse: mtomResponse,
+          gateway: outboundDrRequest.gateway,
+          outboundRequest: {
+            ...outboundDrRequest,
+            documentReference: outboundDrRequest.documentReference.map(docRef => ({
+              ...docRef,
+              metriportId: uuidv4(),
+            })),
+          },
+        },
+      });
+      expect(response?.documentReference?.[0]?.contentType).toEqual(mimeType);
+    });
+  });
+
+  it("[mtom with multiple attachments]: should correctly create and parse MTOM message with XML, PDF, and TIFF attachments", async () => {
+    const xmlFile = testFiles.find(file => file.fileExtension === ".xml");
+    const pdfFile = testFiles.find(file => file.fileExtension === ".pdf");
+    const tiffFile = testFiles.find(file => file.fileExtension === ".tiff");
+
+    if (!xmlFile || !pdfFile || !tiffFile) {
+      throw new Error("Required test files not found");
+    }
+
+    const xmlFilePath = path.join(__dirname, "./files/", xmlFile.name);
+    const pdfFilePath = path.join(__dirname, "./files/", pdfFile.name);
+    const tiffFilePath = path.join(__dirname, "./files/", tiffFile.name);
+
+    const xmlBuffer = fs.readFileSync(xmlFilePath);
+    const pdfBuffer = fs.readFileSync(pdfFilePath);
+    const tiffBuffer = fs.readFileSync(tiffFilePath);
+
+    const attachmentsData = [
+      { payload: xmlBuffer, mimeType: xmlFile.mimeType },
+      { payload: pdfBuffer, mimeType: pdfFile.mimeType },
+      { payload: tiffBuffer, mimeType: tiffFile.mimeType },
+    ];
+
+    const mtomAttachments = await createMtomMessageWithAttachments(attachmentsData);
+    const response = await processDrResponse({
+      drResponse: {
+        mtomResponse: mtomAttachments,
+        gateway: outboundDrRequestMtom.gateway,
+        outboundRequest: outboundDrRequestMtom,
+      },
+    });
+
+    expect(response).toBeDefined();
+    expect(response.documentReference).toHaveLength(3);
+    expect(response.documentReference?.[0]?.contentType).toBe(xmlFile.mimeType);
+    expect(response.documentReference?.[1]?.contentType).toBe(pdfFile.mimeType);
+    expect(response.documentReference?.[2]?.contentType).toBe(tiffFile.mimeType);
+  });
+});
 describe("processDRResponse", () => {
   beforeEach(() => {
     jest.spyOn(S3Utils.prototype, "uploadFile").mockImplementation(() =>
@@ -26,7 +136,7 @@ describe("processDRResponse", () => {
 
   it("should process multiple DR responses correctly", async () => {
     const xmlString = fs.readFileSync(path.join(__dirname, "xmls/dr_success.xml"), "utf8");
-    const mtomResponse = await prepateMtomAttachments(xmlString);
+    const mtomResponse = await createMtomMessageWithoutAttachments(xmlString);
     const response = await processDrResponse({
       drResponse: {
         mtomResponse: mtomResponse,
@@ -52,7 +162,7 @@ describe("processDRResponse", () => {
 
   it("should process the soap fault DR response correctly", async () => {
     const xmlString = fs.readFileSync(path.join(__dirname, "xmls/dr_soap_error.xml"), "utf8");
-    const mtomResponse = await prepateMtomAttachments(xmlString);
+    const mtomResponse = await createMtomMessageWithoutAttachments(xmlString);
     const response = await processDrResponse({
       drResponse: {
         mtomResponse: mtomResponse,
@@ -66,7 +176,7 @@ describe("processDRResponse", () => {
 
   it("should process the registry error DR response correctly", async () => {
     const xmlString = fs.readFileSync(path.join(__dirname, "xmls/dr_registry_error.xml"), "utf8");
-    const mtomResponse = await prepateMtomAttachments(xmlString);
+    const mtomResponse = await createMtomMessageWithoutAttachments(xmlString);
     const response = await processDrResponse({
       drResponse: {
         mtomResponse: mtomResponse,
@@ -79,7 +189,7 @@ describe("processDRResponse", () => {
 
   it("should process the empty DR response correctly", async () => {
     const xmlString = fs.readFileSync(path.join(__dirname, "xmls/dr_empty.xml"), "utf8");
-    const mtomResponse = await prepateMtomAttachments(xmlString);
+    const mtomResponse = await createMtomMessageWithoutAttachments(xmlString);
     const response = await processDrResponse({
       drResponse: {
         mtomResponse: mtomResponse,
@@ -91,7 +201,7 @@ describe("processDRResponse", () => {
   });
   it("should process response that is not a string correctly", async () => {
     const randomResponse = "This is a bad response and is not xml";
-    const mtomResponse = await prepateMtomAttachments(randomResponse);
+    const mtomResponse = await createMtomMessageWithoutAttachments(randomResponse);
     const response = await processDrResponse({
       drResponse: {
         mtomResponse: mtomResponse,
@@ -100,90 +210,5 @@ describe("processDRResponse", () => {
       },
     });
     expect(response.operationOutcome?.issue[0]?.severity).toEqual("information");
-  });
-});
-
-describe.skip("processDRResponse for various file types and verify successful upload without corruption", () => {
-  const s3Utils = new S3Utils(Config.getAWSRegion());
-
-  testFiles.forEach(({ name, mimeType, fileExtension }) => {
-    const xmlTemplatePath = path.join(__dirname, "./xmls/dr-no-mime-type.xml");
-    const xmlTemplate = fs.readFileSync(xmlTemplatePath, "utf8");
-
-    describe(`${name}`, () => {
-      class TestContext {
-        fileContent: Buffer;
-        modifiedXml: string;
-        response?: OutboundDocumentRetrievalResp;
-        key: string;
-        bucket: string | null | undefined;
-        shouldVerifyUpload: boolean;
-
-        constructor(fileName: string) {
-          this.fileContent = fs.readFileSync(path.join(__dirname, `./files/${fileName}`));
-          const fileContentB64 = this.fileContent.toString("base64");
-          this.modifiedXml = xmlTemplate.replace(
-            "<Document></Document>",
-            `<Document>${fileContentB64}</Document>`
-          );
-          this.key = "";
-          this.bucket = "";
-          this.shouldVerifyUpload = testFilesForUploadVerification.some(
-            file => file.name === fileName
-          );
-        }
-
-        async processDRResponse() {
-          const mtomResponse = await prepateMtomAttachments(this.modifiedXml);
-          this.response = await processDrResponse({
-            drResponse: {
-              mtomResponse: mtomResponse,
-              gateway: outboundDrRequest.gateway,
-              outboundRequest: {
-                ...outboundDrRequest,
-                documentReference: outboundDrRequest.documentReference.map(docRef => ({
-                  ...docRef,
-                  metriportId: uuidv4(),
-                })),
-              },
-            },
-          });
-
-          this.key = this.response.documentReference?.[0]?.fileName || "";
-          this.bucket = this.response.documentReference?.[0]?.fileLocation;
-        }
-
-        async verifyUpload() {
-          if (!this.bucket || !this.shouldVerifyUpload) {
-            return;
-          }
-          const fileInfo = await s3Utils.getFileInfoFromS3(this.key, this.bucket);
-          expect(fileInfo.contentType).toEqual(mimeType);
-
-          const downloadedContent = await s3Utils.downloadFile({
-            bucket: this.bucket,
-            key: this.key,
-          });
-          expect(downloadedContent).toEqual(this.fileContent);
-
-          await s3Utils.deleteFile({ bucket: this.bucket, key: this.key });
-        }
-      }
-
-      let testContext: TestContext;
-
-      beforeAll(async () => {
-        testContext = new TestContext(name);
-        await testContext.processDRResponse();
-      }, 20000);
-
-      it(`should process the ${fileExtension} DR response correctly`, () => {
-        expect(testContext.response?.documentReference?.[0]?.contentType).toEqual(mimeType);
-      });
-
-      it(`should verify the ${fileExtension} file upload to S3 without corruption if applicable`, async () => {
-        await testContext.verifyUpload();
-      }, 20000);
-    });
   });
 });
